Split FindCommandTest into FindPersonCommandTest and FindOwnerCommandTest
