+++ src/test/java/seedu/address/logic/commands/FindPersonCommandTest.java
@@ -4,7 +4,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.logic.Messages.MESSAGE_PERSONS_LISTED_OVERVIEW;
-import static seedu.address.logic.Messages.MESSAGE_OWNERS_LISTED_OVERVIEW;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.testutil.TypicalPersons.CARL;
 import static seedu.address.testutil.TypicalPersons.ELLE;
@@ -18,20 +17,15 @@ import org.junit.jupiter.api.Test;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.owner.OwnerNameContainsKeywordsPredicate;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
-import seedu.address.testutil.TypicalOwners;
 import seedu.address.testutil.TypicalPersons;
 
 /**
- * Contains integration tests (interaction with the Model) for {@code FindCommand}.
+ * Contains integration tests (interaction with the Model) for {@code FindPersonCommand}.
  */
-public class FindCommandTest {
+public class FindPersonCommandTest {
     private Model model = new ModelManager(TypicalPersons.getTypicalAddressBook(), new UserPrefs());
     private Model expectedModel = new ModelManager(TypicalPersons.getTypicalAddressBook(), new UserPrefs());
-
-    private Model ownerModel = new ModelManager(TypicalOwners.getTypicalAddressBook(), new UserPrefs());
-    private Model expectedOwnerModel = new ModelManager(TypicalOwners.getTypicalAddressBook(), new UserPrefs());
 
     @Test
     public void equals() {
@@ -59,31 +53,6 @@ public class FindCommandTest {
 
         // different person -> returns false
         assertFalse(findFirstCommand.equals(findSecondCommand));
-
-        // Test for owners
-        OwnerNameContainsKeywordsPredicate firstOwnerPredicate =
-                new OwnerNameContainsKeywordsPredicate(Collections.singletonList("first"));
-        OwnerNameContainsKeywordsPredicate secondOwnerPredicate =
-                new OwnerNameContainsKeywordsPredicate(Collections.singletonList("second"));
-
-        FindOwnerCommand findOwnerFirstCommand = new FindOwnerCommand(firstOwnerPredicate);
-        FindOwnerCommand findOwnerSecondCommand = new FindOwnerCommand(secondOwnerPredicate);
-
-        // same object -> returns true
-        assertTrue(findOwnerFirstCommand.equals(findOwnerFirstCommand));
-
-        // same values -> returns true
-        FindOwnerCommand findOwnerFirstCommandCopy = new FindOwnerCommand(firstOwnerPredicate);
-        assertTrue(findOwnerFirstCommand.equals(findOwnerFirstCommandCopy));
-
-        // different types -> returns false
-        assertFalse(findOwnerFirstCommand.equals(1));
-
-        // null -> returns false
-        assertFalse(findOwnerFirstCommand.equals(null));
-
-        // different owners -> returns false
-        assertFalse(findOwnerFirstCommand.equals(findOwnerSecondCommand));
     }
 
     @Test
@@ -97,16 +66,6 @@ public class FindCommandTest {
     }
 
     @Test
-    public void execute_zeroKeywords_noOwnerFound() {
-        String expectedMessage = String.format(MESSAGE_PERSONS_LISTED_OVERVIEW, 0);
-        OwnerNameContainsKeywordsPredicate predicate = prepareOwnerPredicate(" ");
-        FindOwnerCommand command = new FindOwnerCommand(predicate);
-        expectedOwnerModel.updateFilteredOwnerList(predicate);
-        assertCommandSuccess(command, ownerModel, expectedMessage, expectedOwnerModel);
-        assertEquals(Collections.emptyList(), ownerModel.getFilteredOwnerList());
-    }
-
-    @Test
     public void execute_multipleKeywords_multiplePersonsFound() {
         String expectedMessage = String.format(MESSAGE_PERSONS_LISTED_OVERVIEW, 3);
         NameContainsKeywordsPredicate predicate = preparePredicate("Kurz Elle Kunz");
@@ -114,19 +73,6 @@ public class FindCommandTest {
         expectedModel.updateFilteredPersonList(predicate);
         assertCommandSuccess(command, model, expectedMessage, expectedModel);
         assertEquals(Arrays.asList(CARL, ELLE, FIONA), model.getFilteredPersonList());
-    }
-
-    @Test
-    public void execute_multipleKeywords_multipleOwnersFound() {
-        String expectedOwnerMessage = String.format(MESSAGE_OWNERS_LISTED_OVERVIEW, 3);
-        OwnerNameContainsKeywordsPredicate ownerPredicate = prepareOwnerPredicate("Kurz Elle Kunz");
-        FindOwnerCommand commandOwner = new FindOwnerCommand(ownerPredicate);
-        expectedOwnerModel.updateFilteredOwnerList(ownerPredicate);
-        assertCommandSuccess(commandOwner, ownerModel, expectedOwnerMessage, expectedOwnerModel);
-        assertEquals(Arrays.asList(seedu.address.testutil.TypicalOwners.CARL,
-                        seedu.address.testutil.TypicalOwners.ELLE,
-                        seedu.address.testutil.TypicalOwners.FIONA),
-                ownerModel.getFilteredOwnerList());
     }
 
     @Test
@@ -142,12 +88,5 @@ public class FindCommandTest {
      */
     private NameContainsKeywordsPredicate preparePredicate(String userInput) {
         return new NameContainsKeywordsPredicate(Arrays.asList(userInput.split("\\s+")));
-    }
-
-    /**
-     * Parses {@code userInput} into a {@code OwnerNameContainsKeywordsPredicate}.
-     */
-    private OwnerNameContainsKeywordsPredicate prepareOwnerPredicate(String userInput) {
-        return new OwnerNameContainsKeywordsPredicate(Arrays.asList(userInput.split("\\s+")));
     }
 }
